third_party/blink/web_tests/platform/win11-arm64/external/wpt/visual-viewport/page-and-offset-in-iframe-expected.txt

This is a testharness.js-based test.
[FAIL] VisualViewport page and offset values in iframe
  assert_greater_than: Pinch zoom must have increased scale expected a number greater than 1.2 but got 1
Harness: the test ran to completion.
